requirements.txt
certifi==2021.5.30
charset-normalizer==2.0.3
click==8.0.1
colorama==0.4.4
Flask==2.0.1
future==0.18.2
gunicorn==20.1.0
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
joblib==1.0.1
line-bot-sdk==1.19.0
MarkupSafe==2.0.1
nltk==3.6.2
numpy==1.21.1
pandas==1.3.1
python-dateutil==2.8.2
pytz==2021.1
regex==2021.7.6
requests==2.26.0
Sastrawi==1.0.1
six==1.16.0
tqdm==4.61.2
urllib3==1.26.6
Werkzeug==2.0.1
